reduce buildtests_cxx target to end2end tests
as of v1.69, the whole set of tests takes extremely long to build (>6h),
so reduce it to only "end2end" tests. This has the advantage of being
easily greppable (to make a choice of subset, and an easy rule for rebasing),
as well as being the most high-level tests that assure the library functions
as intended.

Plus four support libraries (`grpc{,++}_test{,_config,_util}`) and the vendored gtest