reduce grpc symbol explosion with autogenerated export definitions file
Basically, WINDOWS_EXPORT_ALL_SYMBOLS is a nuclear option that includes all symbols
that occur anywhere in the various object files into an export.

In doing so, it mixes two fundamental concerns:
* building the project itself (without having to add lots of macros that do the __declspec dance).
* visibility of symbols from transitive dependencies (like abseil, openssl, etc.) even if these are
  marked as being linked privately.

To give some perspective, the python pefile library shows roughly 10k named/numbered symbols in
grpc.dll, and yet as of v1.68, it blew through the limit of 65535 symbols, which is the maximum
that DLLs can contain. So roughly 5/6th of the symbols are garbage.

In order to make this salvage this situation, we turn off WINDOWS_EXPORT_ALL_SYMBOLS (no other choice)
and add a PRE_LINK script that analyses the symbols in the various object files that have been compiled
for the grpc target. This is based on the "source of truth", i.e. `dumpbin /symbols`, which is part of
the MSVC toolchain. We need to do some light parsing of these dumps, and then decide which symbols to
export explicitly, using the `/DEF:grpc.def` linker command.

Here we need to thread the needle in the sense that we can only put stuff in the `grpc.def` file that
the linker can actually legally find. For example, `Static` symbols that are internal to the objects
would not be found, if we included them. At the same time, the export must be comprehensive enough
that all consumers of the libraries (here mainly grpc++.dll) find the symbols they expect. For this
to work, we need to mark a couple of extra symbols as being exported, using the GRPC_DLL mechanism.

The script that does this (generate_def.py) does some minimal munging and filtering, which could
certainly be sharpened based on a deeper understanding of the Itanium mangling. For now, we do not
filter as aggressively as possible (e.g. lambdas), because our main goal is to get below the 65k
limit, and indeed, we end up <25k, and we don't want to remove symbols over-zealously if some
other consumers end up linking to them.